requirements.txt
pandas
scikit-learn
numpy
joblib
fastapi
uvicorn[standard]
pydantic
openai
python-dotenv
